refactor(inspection-edit): extract photo/media ops into usePhotoOps hook (2237->1817)
Batch 4: moves the photo/media-operations cluster (viewer/crop/recrop/poster/
gallery-crop UI state + photoOpsFetcher + videoPosterUrl/getItemPhotos/
itemGalleryPhotos/onOpenPhoto/patchItemPhotos/onReorderPhotos/onBulkDetachPhotos/
moveTargets/onBulkMovePhotos/onViewerAction/performPhotoCropSave) into
app/hooks/usePhotoOps.ts, mirroring the useInspection/useFindings ctx-in
object-out pattern. coverFetcher stays in the component (shared with non-photo
overlays). Destructured into identical local names so all overlay JSX is unchanged.

Behavior-preserving: reviewer diffed the 319-line moved region -> byte-identical;
all 10 useCallback/useMemo dependency arrays unchanged; React hook-order
preserved. Gates: type-check 0, lint 0, test:web 516, db:check clean.
